af/requirements.txt
apache-airflow-providers-google
pyarrow
pandas
sqlalchemy
psycopg2-binary
dbt-bigquery==1.3.0